feat: Add AppDbContext for managing user attendance
This commit adds the `AppDbContext` class to the `Infrastructure/Data` directory. The `AppDbContext` class extends `DbContext` and includes `DbSet` properties for `User` and `Attendance` entities. It also configures the relationships between the entities using Fluent API.

The `AppDbContext` will be used to manage user attendance data and provide access to the `Users` and `Attendances` tables in the database.