refactor: Rename User class to Employee and update property names
The User class in the Core/Entities directory has been renamed to Employee. The property "Name" has been changed to "FirstName" and a new property "FullName" has been added to concatenate the first and last names. This change provides a more accurate representation of the entity and improves the clarity of the code.